Module mail_internals::grammar::encoded_word

Grammar parts for encoded words (based on RFC 2047).

Enums

EncodedWordContext
Represents the place at which the encoded word appears.

Constants

ECW_SEP_OVERHEAD
The syntax overhead from "framing" an encoded word.

MAX_ECW_LEN
maximal length of an encoded word

Functions

is_encoded_word
Returns true if the given word is a encoded word.

try_parse_encoded_word_parts
Tries to parse the given string as an encoded word.