Function mail_internals::grammar::encoded_word::is_encoded_word

pub fn is_encoded_word(
    word: &str, 
    ctx: EncodedWordContext, 
    mail_type: MailType
) -> bool

Returns true if the given word is a encoded word.

Note that this depends on the context the word appears in and the mail type. The reason for this is that encoded words tend to be valid text even without decoding them. But this means if the encoded word has some syntax error (e.g. missing closing ?=) it is no longer an encoded word but just some text which happen to look similar to one.