Module mail_internals::utils

Some more general utilities.

Or with other words, thinks which (currently) have no other place to be placed in.

Structs

DebugIterableOpaque
Helper for implementing debug for an iterable think where the think on itself is irrelevant.

Functions

is_utf8_continuation_byte
returns true if this a not first byte from a multi byte utf-8

uneraser_ref
Used to undo type erasure in a generic context, roughly semantically eqivalent to creating a &Any type object from the input and then using downcast_ref::<EXP>(), except that it does not require the cration of a trait object as a step inbetween.

vec_insert_bytes
Faster insertion of byte slices into a byte vector.